refactor: form 컴포넌트 안에 있던 form-tab-manager을 page컴포넌트로 이동 #74
