+++ src/app/concert/form/[id]/page.tsx
@@ -2,12 +2,15 @@
 import { use } from 'react';
 
 import AppBarSetter from '@/shared/components/header/app-bar/app-bar-setter';
+import { useModal } from '@/shared/components/modal/use-modal';
 import { customToast } from '@/shared/components/toast/custom-toast/custom-toast';
 import { TicketOpenType } from '@/shared/types';
 
 import Form from './_shared/components/form/form';
+import FormModal from './_shared/components/form-modal/form-modal';
 import styles from './page.module.scss';
 import { useGetConcertDetail } from '../../[id]/_shared/services/query';
+import FormTabManager from './_shared/components/tab-button/manager/form-tab-manager';
 
 export default function Page({
   params,
@@ -18,8 +21,8 @@ export default function Page({
 }) {
   const resolvedParams = use(params);
   const resolvedSearchParams = use(searchParams);
-
   const { id } = resolvedParams;
+  const { open, closeTop } = useModal();
 
   const ticketOpenType = resolvedSearchParams.ticketOpenType as TicketOpenType;
 
@@ -32,18 +35,46 @@ export default function Page({
       description: message,
     });
 
+  const handleOpenModal = () => {
+    open({
+      id: 'form-modal',
+      content: (
+        <FormModal
+          title="일반예매 신청이 완료되었습니다."
+          message={`대리인이 수락하게 되면 매칭이 완료됩니다.\n매칭이 완료되면 채팅을 통해 이야기를 나눠보세요.`}
+          onConfirm={async () => {
+            await new Promise((resolve) => setTimeout(resolve, 1000));
+            closeTop();
+          }}
+          onCancel={() => {
+            closeTop();
+          }}
+          concertId={id}
+        />
+      ),
+    });
+  };
+
   return (
     <>
       <AppBarSetter title="신청 양식" />
 
       <div className={styles.container}>
         {concertItem && (
-          <Form
-            concertItem={concertItem}
-            ticketOpenType={ticketOpenType}
-            concertId={id}
-            onError={handleErrorToast}
-          />
+          <>
+            <Form
+              concertItem={concertItem}
+              ticketOpenType={ticketOpenType}
+              concertId={id}
+            />
+            <FormTabManager
+              handleOpenModal={handleOpenModal}
+              concertItem={concertItem}
+              ticketOpenType={ticketOpenType}
+              concertId={id}
+              onError={handleErrorToast}
+            />
+          </>
         )}
       </div>
     </>
