+++ src/shared/types/concert.ts
@@ -56,4 +56,5 @@ export type {
   Concert,
   TicketOpenType,
   TicketOpenDateInfo,
+  ConcertDateInfo,
 };
